reorganize build
